\name{archery2}
\alias{archery2}
\docType{data}
\title{
Target archery dataset during the elimination stage (used as Phase II)
}
\description{
It consists of a stage in which the archer shoots 52 arrows.
The information is given in x and y coordinates. 
}
\usage{data(archery1)}
\format{
An array of (18 x 2 x 3).
  \describe{
    \item{\code{"x-coordinate"}}{x coordinate}
    \item{\code{"y-coordinate"}}{y coordinate}
  }
}
\examples{
data(archery1)
## maybe str(archery1) ; plot(archery1) ...
}
\keyword{datasets}