\name{dowel1}
\alias{dowel1}
\docType{data}
\title{
Dowel pin dataset in Phase I
}
\description{
Diameter and length of a manufacturing process of a dowel pin
}
\usage{data(dowel1)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{diameter}}{a numeric vector}
    \item{\code{length}}{a numeric vector}
  }
}

\examples{
data(dowel1)
## maybe str(dowel1) ; plot(dowel1) ...
}
\keyword{datasets}
