\name{dowel2}
\alias{dowel2}
\docType{data}
\title{
Dowel pin dataset in Phase II
}
\description{
Diameter and length of a manufacturing process of a dowel pin
}
\usage{data(dowel2)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{diameter}}{a numeric vector}
    \item{\code{length}}{a numeric vector}
  }
}

\examples{
data(dowel2)
## maybe str(dowel2) ; plot(dowel2) ...
}
\keyword{datasets}
