\name{glass1}
\alias{glass1}
\docType{data}
\title{
Glass manufacturing in Phase I
}
\description{
It contains three variables measured with the aim to establish a multivariate monitoring program in a manufacturing process
}
\usage{data(glass1)}
\format{
An array of (32 x 3 x 5).
  \describe{
    \item{\code{"Var1"}}{a numeric vector}
    \item{\code{"Var2"}}{a numeric vector }
	\item{\code{"Var3"}}{a numeric vector }
  }
}

\examples{
data(glass1)
## maybe str(glass1) ; plot(glass1) ...
}
\keyword{datasets}
