\name{glass2}
\alias{glass2}
\docType{data}
\title{
Glass manufacturing in Phase II
}
\description{
It contains three variables measured with the aim to establish a multivariate monitoring program in a manufacturing process
}
\usage{data(glass2)}
\format{
An array of (32 x 3 x 5).
  \describe{
    \item{\code{"Var1"}}{a numeric vector}
    \item{\code{"Var2"}}{a numeric vector }
	\item{\code{"Var3"}}{a numeric vector }
  }
}

\examples{
data(glass2)
## maybe str(glass2) ; plot(glass2) ...
}
\keyword{datasets}
