\name{indust1}
\alias{indust1}
\docType{data}
\title{
Industrial dataset collected in Phase I
}
\description{
A bivariate industrial process 
}
\usage{data(indust1)}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{Var1}}{a numeric vector}
    \item{\code{Var2}}{a numeric vector}
  }
}

\examples{
data(indust1)
## maybe str(indust1) ; plot(indust1) ...
}
\keyword{datasets}
