\name{indust2}
\alias{indust2}
\docType{data}
\title{
Industrial dataset collected in Phase II
}
\description{
A bivariate industrial process 
}
\usage{data(indust2)}
\format{
  A data frame containing 35 observations.
  \describe{
    \item{\code{Var1}}{a numeric vector}
    \item{\code{Var2}}{a numeric vector}
  }
}

\examples{
data(indust2)
## maybe str(indust2) ; plot(indust2) ...
}
\keyword{datasets}
