\name{mech1}
\alias{mech1}
\docType{data}
\title{
A mechanical process in Phase I
}
\description{
Seven variables collected from a mechanical process
}
\usage{data(mech1)}
\format{
The format is:
An array of (45 x 7).
 num [1:45, 1:7] 9.55 9.95 10.5 8.75 9.98 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "var1" "var2" "var3" "var4" ...
}

\examples{
data(mech1)
## maybe str(mech1) ; plot(mech1) ...
}
\keyword{datasets}
