\name{mech2}
\alias{mech2}
\docType{data}
\title{
A mechanical process in Phase II
}
\description{
Seven variables collected from a mechanical process
}
\usage{data(mech2)}
\format{
  The format is:
  An array of (50 x 7).
 num [1:50, 1:7] 9.95 10.08 9.19 10.51 9.39 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "var1" "var2" "var3" "var4" ...
}

\examples{
data(mech2)
## maybe str(mech2) ; plot(mech2) ...
}
\keyword{datasets}
