\name{rskewed}
\alias{rskewed}
\docType{data}
\title{
Right skewed dataset
}
\description{
It is a bivariate dataset with a presence of right skewness
}
\usage{data(rskewed)}
\format{
  The format is:
 num [1:30, 1:2] 0.763 0.546 0.787 0.535 0.519 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "var1" "var2"
}
\examples{
data(rskewed)
## maybe str(rskewed) ; plot(rskewed) ...
}
\keyword{datasets}
