\name{sabathia.ind}
\alias{sabathia.ind}
\docType{data}
\title{
Individual observations from the MLB pitcher C.C. Sabathia from the game on July 10, 2011. 
}
\description{
The dataset was selected from the game against Tampa Bay on July 10, 2011. It is composed by individual observations.
}
\usage{data(sabathia.ind)}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{px}}{is the x-axis horizontally oriented}
    \item{\code{pz}}{is the z-axis vertically oriented}
  }
}
\examples{
data(sabathia.ind)
}
\keyword{datasets}
