\name{sabathia1}
\alias{sabathia1}
\docType{data}
\title{
A pitching log of C.C. Sabathia on July 10, 2011.
}
\description{
The dataset is from the game against Tampa Bay on July 10, 2011. It contains the mean of the rational subgroup.
}
\usage{data(sabathia1)}
\format{
  A data frame with 23 observations.
  \describe{
    \item{\code{px}}{the x-axis horizontally oriented}
    \item{\code{pz}}{the z-axis vertically oriented }
    \item{\code{start speed}}{is the starting speed of the fastball}
  }
}
\examples{
data(sabathia1)
}
\keyword{datasets}
