\name{sabathia2}
\alias{sabathia2}
\docType{data}
\title{
A pitching log of C.C. Sabathia on August 12, 2011
}
\description{
The dataset was selected from games against Tampa Bay on August 12, 2011 
}
\usage{data(sabathia2)}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{px}}{the x-axis is the horizontally oriented}
    \item{\code{pz}}{the z-axis is the vertically oriented }
    \item{\code{start speed}}{is the start speed of the fastball}
  }
}

\examples{
data(sabathia2)
## maybe str(sabathia2) ; plot(sabathia2) ...
}
\keyword{datasets}
