\name{water2}
\alias{water2}
\docType{data}
\title{
A water quality test in Phase II
}
\description{
It consists on five variables (pH, phosphates (mg/L), nitrates (mg/L), dissolved oxygen and total solids (mg/L)) measured in a water quality test
}
\usage{data(water2)}
\format{
  The format is:
 is a matrix (25 x 5)
}

\examples{
data(water2)
## maybe str(water2) ; plot(water2) ...
}
\keyword{datasets}
