\name{carbon2}
\alias{carbon2}
\docType{data}
\title{
Carbon fiber tubing during the Phase II
}
\description{
Three quality characteristics (inner diameter, thickness and length) are measured in a specific carbon fiber tube. 
This dataset consist of 25 samples of size 8. This dataset is considered as the Phase II.
}
\usage{data(carbon2)}
\format{
An array of dimensions 25 x 3 x 8.
  \describe{
    \item{\code{"inner diameter"}}{is the inner diameter}
    \item{\code{"thickness"}}{is the thickness}
	\item{\code{"length"}}{is the length}
  }
}

\examples{
data(carbon2)
## maybe str(carbon2) ; plot(carbon2) ...
}
\keyword{datasets}
