\name{dowel2}
\alias{dowel2}
\docType{data}
\title{
Dowel pin dataset for the Phase II
}
\description{
Diameter and length of a dowel pin
}
\usage{data(dowel2)}
\format{
  A data frame with 32 observations.
  \describe{
    \item{\code{diameter}}{a numeric vector}
    \item{\code{length}}{a numeric vector}
  }
}

\examples{
data(dowel2)
## maybe str(dowel2) ; plot(dowel2) ...
}
\keyword{datasets}
