\name{glass2}
\alias{glass2}
\docType{data}
\title{
Dataset of a glass manufacturing process during the Phase II
}
\description{
It contains the measurements of three quality characteristics
}
\usage{data(glass2)}
\format{
An array of (25 x 3 x 5).
  \describe{
    \item{\code{"Var1"}}{a numeric vector}
    \item{\code{"Var2"}}{a numeric vector}
	\item{\code{"Var3"}}{a numeric vector}
  }
}

\examples{
data(glass2)
## maybe str(glass2) ; plot(glass2) ...
}
\keyword{datasets}
