\name{kulpa}
\alias{kulpa}
\docType{data}
\title{
Coordinates of the pitches called "strike" by umpire Ron Kulpa
}
\description{
The dataset was selected from a Tampa Bay game on July 10, 2011 
}
\usage{data(kulpa)}
\format{
  A data frame with 113 observations on the following 2 variables.
  \describe{
    \item{\code{px}}{the x-axis is horizontally oriented}
    \item{\code{pz}}{the z-axis is the vertically oriented}
  }
}
\examples{
data(kulpa)
}
\keyword{datasets}
