\name{sabathia.ind}
\alias{sabathia.ind}
\docType{data}
\title{
Individual observations of C.C. Sabathia on July 10, 2011. 
}
\description{
The dataset was selected from games against Tampa Bay on July 10, 2011. It is composed by the individual observations.
}
\usage{data(sabathia.ind)}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{px}}{the x-axis is the horizontally oriented}
    \item{\code{pz}}{the z-axis is the vertically oriented }
  }
}
\examples{
data(sabathia.ind)
}
\keyword{datasets}
