\name{sabathia1}
\alias{sabathia1}
\docType{data}
\title{
A pitching log of C.C. Sabathia on July 10, 2011.
}
\description{
The dataset was selected from games against Tampa Bay on July 10, 2011. It is composed by the mean of the rational subgroup.
}
\usage{data(sabathia1)}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{px}}{the x-axis is the horizontally oriented}
    \item{\code{pz}}{the z-axis is the vertically oriented }
    \item{\code{start speed}}{is the start speed of the fastball}
  }
}
\examples{
data(sabathia1)
}
\keyword{datasets}
