\name{sabathia2}
\alias{sabathia2}
\docType{data}
\title{
A pitching log of C.C. Sabathia on August 12, 2011
}
\description{
The dataset is from the game against Tampa Bay on August 12, 2011 
}
\usage{data(sabathia2)}
\format{
  A data frame with 26 observations.
  \describe{
    \item{\code{px}}{the x-axis horizontally oriented}
    \item{\code{pz}}{the z-axis vertically oriented }
    \item{\code{start speed}}{is the starting speed of the fastball}
  }
}

\examples{
data(sabathia2)
}
\keyword{datasets}
