\name{water1}
\alias{water1}
\docType{data}
\title{
A water quality test in Phase I
}
\description{
It consists on five variables (pH, phosphates (mg/L), nitrates (mg/L), dissolved oxygen and total solids (mg/L)) measured in a water quality test
}
\usage{data(water1)}
\format{
  The format is:
is a matrix (30 x 5)
}

\examples{
data(water1)
## maybe str(water1) ; plot(water1) ...
}
\keyword{datasets}
